refactor: 함수 쪼개기
