--- src/hooks/cardNumbers/useCardNumbersRef.ts
@@ -1,12 +1,0 @@
-import { useRef } from 'react';
-
-export const useCardNumbersRef = () => {
-  const inputRefs = {
-    first: useRef<HTMLInputElement>(null),
-    second: useRef<HTMLInputElement>(null),
-    third: useRef<HTMLInputElement>(null),
-    fourth: useRef<HTMLInputElement>(null)
-  };
-
-  return { inputRefs };
-};
