+++ src/hooks/cardNumbers/useCardNumbersState.ts
@@ -1,15 +1,18 @@
-import { useState } from 'react';
+import { useRef, useState } from 'react';
 import { CardNumberType } from '../../types';
 import { INITIAL_CARD_NUMBER } from '../../constants';
 import { isNumber } from '../../utils/isNumber';
-import { validateCardNumber } from '../../validation/validateCardNumbers';
 import focusNextInputIfFilled from '../../utils/focusNextInputIfFilled';
-import { useCardNumbersRef } from './useCardNumbersRef';
+import { isValidCardNumber, isValidCardNumbers } from '../../validation/validateCardNumbers';
 
 export const useCardNumbersState = () => {
   const [cardNumbers, setCardNumbers] = useState<CardNumberType>(INITIAL_CARD_NUMBER);
-  const { isValid } = validateCardNumber(cardNumbers);
-  const { inputRefs } = useCardNumbersRef();
+  const inputRefs = {
+    first: useRef<HTMLInputElement>(null),
+    second: useRef<HTMLInputElement>(null),
+    third: useRef<HTMLInputElement>(null),
+    fourth: useRef<HTMLInputElement>(null)
+  };
 
   const handleCardNumbersChange = (field: keyof CardNumberType, value: string) => {
     if (value !== '' && !isNumber(value)) {
@@ -17,7 +20,7 @@ export const useCardNumbersState = () => {
     }
     setCardNumbers((prev) => ({
       ...prev,
-      [field]: { value, isError: isValid }
+      [field]: { value, isError: !isValidCardNumber(value) }
     }));
 
     const keys = Object.keys(inputRefs);
